backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kickoff"
version = "0.1.0"
description = "Kickoff — Tournament management platform"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<6.0",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "PyJWT>=2.8,<3.0",
    "drf-spectacular>=0.28,<1.0",
    "django-cors-headers>=4.3,<5.0",
    "django-filter>=24.0",
    "Pillow>=10.0",
    "qrcode[pil]>=8.0,<9.0",
    "psycopg2-binary>=2.9,<3.0",
    "python-decouple>=3.8",
    # Channels (WebSocket)
    "channels>=4.1,<5.0",
    "channels-redis>=4.2,<5.0",
    "daphne>=4.1,<5.0",
    # Celery (task queue)
    "celery>=5.4,<6.0",
    "redis>=5.0,<6.0",
    # Monitoring
    "sentry-sdk[django]>=2.0,<3.0",
    "django-structlog>=8.0,<9.0",
    # Rate limiting
    "django-ratelimit>=4.1,<5.0",
    # Production
    "gunicorn>=22.0,<23.0",
    "whitenoise>=6.7,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "ruff>=0.8",
    "mypy>=1.13",
    "django-stubs>=5.1",
    "djangorestframework-stubs>=3.15",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["."]
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"kickoff/settings/dev.py" = ["F401", "F403", "F405"]
"kickoff/settings/prod.py" = ["F401", "F403", "F405"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = false
warn_unused_configs = true
warn_redundant_casts = true

[tool.django-stubs]
django_settings_module = "kickoff.settings.dev"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "kickoff.settings.dev"
python_files = ["test_*.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
